change money to decimal

+++ inventory/models/Product.py
@@ -3,7 +3,7 @@ from inventory.models.UnitOfMeasure import UnitOfMeasure
 
 class Product(models.Model):
 	description = models.CharField(max_length=255)
-	sell_price = models.FloatField(default=0)
+	sell_price = models.DecimalField(max_digits=20, decimal_places=2, default=0.00)
 	show_in_menu= models.BooleanField()
 	is_raw_material= models.BooleanField()
 	is_finish_product= models.BooleanField(default=0)
